[SignIn] page configured as container pattern
