+++ client/src/components/templates/SignIn/index.tsx
@@ -1,15 +1,3 @@
-import * as AppleAuthentication from 'expo-apple-authentication';
-import * as Crypto from 'expo-crypto';
-import * as Device from 'expo-device';
-
-import {
-  Alert,
-  Dimensions,
-  Image,
-  Platform,
-  TouchableOpacity,
-  View,
-} from 'react-native';
 import Animated, {
   block,
   clockRunning,
@@ -18,8 +6,14 @@ import Animated, {
   set,
   useCode,
 } from 'react-native-reanimated';
-import {AuthPayload, User} from '../../../types/graphql';
 import {Button, EditText, ThemeType, useTheme} from 'dooboo-ui';
+import {
+  Dimensions,
+  Image,
+  Platform,
+  TouchableOpacity,
+  View,
+} from 'react-native';
 import {
   IC_LOGO_D,
   IC_LOGO_W,
@@ -27,29 +21,14 @@ import {
   SvgFacebook,
   SvgGoogle,
 } from '../../../utils/Icons';
-import React, {ReactElement, useEffect, useState} from 'react';
-import type {
-  UserSignInAppleMutation,
-  UserSignInAppleMutationResponse,
-} from '../../../__generated__/UserSignInAppleMutation.graphql';
-import type {
-  UserSignInEmailMutation,
-  UserSignInEmailMutationResponse,
-} from '../../../__generated__/UserSignInEmailMutation.graphql';
+import React, {FC, ReactElement} from 'react';
 import {delay, spring, useClock, useValue} from 'react-native-redash';
-import {showAlertForError, validateEmail} from '../../../utils/common';
-import {signInEmail, signInWithApple} from '../../../relay/queries/User';
 import styled, {css} from 'styled-components/native';
 
-import AsyncStorage from '@react-native-async-storage/async-storage';
-import {AuthStackNavigationProps} from '../../navigations/AuthStackNavigator';
-import type {NotificationCreateNotificationMutation} from '../../../__generated__/NotificationCreateNotificationMutation.graphql';
 import SocialSignInButton from './SocialSignInButton';
 import StatusBar from '../../UI/atoms/StatusBar';
-import {createNotification} from '../../../relay/queries/Notification';
+import {User} from '../../../types/graphql';
 import {getString} from '../../../../STRINGS';
-import {useAuthContext} from '../../../providers/AuthProvider';
-import {useMutation} from 'react-relay/hooks';
 
 const AnimatedTouchableOpacity = Animated.createAnimatedComponent(
   TouchableOpacity,
@@ -131,171 +110,43 @@ const StyledScrollView = styled.ScrollView`
 `;
 
 interface Props {
-  navigation: AuthStackNavigationProps<'SignIn'>;
+  email: string;
+  password: string;
+  errorPassword: string;
+  errorEmail: string;
+  isSigningIn: boolean;
+  isSigningInWithApple: boolean;
+  onAppIconPressed: () => void;
+  onSignInPressed: () => void;
+  onSignUpPressed: () => void;
+  onFindPwPressed: () => void;
+  onSignInWithApplePressed: () => void;
+  onUserCreated: (user?: User) => void;
+  onAgreementPressed: () => void;
+  onPrivacyPressed: () => void;
+  onChangeEmailText: (text: string) => void;
+  onChangePasswordText: (text: string) => void;
 }
 
-function SignIn(props: Props): ReactElement {
-  const {navigation} = props;
-  const {setUser} = useAuthContext();
-  const {theme, changeThemeType, themeType} = useTheme();
-
-  const [signingInApple, setSigningInApple] = useState<boolean>(false);
-  const [email, setEmail] = useState<string>('');
-  const [password, setPassword] = useState<string>('');
-  const [errorEmail, setErrorEmail] = useState<string>('');
-  const [errorPassword, setErrorPassword] = useState<string>('');
-
-  const [commitEmail, isInFlight] = useMutation<UserSignInEmailMutation>(
-    signInEmail,
-  );
-
-  const [commitApple, isAppleInFlight] = useMutation<UserSignInAppleMutation>(
-    signInWithApple,
-  );
-
-  const [
-    commitNotification,
-  ] = useMutation<NotificationCreateNotificationMutation>(createNotification);
-
-  const createNotificationIfPushTokenExists = async (): Promise<void> => {
-    const pushToken = await AsyncStorage.getItem('push_token');
-
-    if (pushToken) {
-      const createNotificationMutationConfig = {
-        variables: {
-          token: pushToken,
-          device: Device.modelName,
-          os: Device.osName,
-        },
-      };
-
-      commitNotification(createNotificationMutationConfig);
-    }
-  };
-
-  const goToSignUp = (): void => {
-    navigation.navigate('SignUp');
-  };
-
-  const goToFindPw = (): void => {
-    navigation.navigate('FindPw');
-  };
-
-  const signIn = async (): Promise<void> => {
-    if (!validateEmail(email)) {
-      setErrorEmail(getString('EMAIL_FORMAT_NOT_VALID'));
-
-      return;
-    }
-
-    if (!password) {
-      setErrorPassword(getString('PASSWORD_REQUIRED'));
-
-      return;
-    }
-
-    const licenseAgreed = JSON.parse(
-      (await AsyncStorage.getItem('license_agreed')) as string,
-    );
-
-    if (!licenseAgreed) return navigation.navigate('LicenseAgreement');
-
-    const mutationConfig = {
-      variables: {
-        email,
-        password,
-      },
-
-      onCompleted: (response: UserSignInEmailMutationResponse) => {
-        const {token, user} = response.signInEmail as AuthPayload;
-
-        if (user && !user.verified)
-          return navigation.navigate('VerifyEmail', {
-            email,
-          });
-
-        AsyncStorage.setItem('token', token);
-
-        createNotificationIfPushTokenExists();
-
-        setUser(user as User);
-      },
-
-      onError: (error: Error): void => {
-        setErrorPassword(error.message);
-        setErrorPassword(getString('PASSWORD_INCORRECT'));
-      },
-    };
-
-    commitEmail(mutationConfig);
-  };
-
-  const goToWebView = (uri: string): void => {
-    props.navigation.navigate('WebView', {uri});
-  };
-
-  const appleLogin = async (): Promise<void> => {
-    setSigningInApple(true);
-
-    try {
-      const csrf = Math.random().toString(36).substring(2, 15);
-      const nonce = Math.random().toString(36).substring(2, 10);
-
-      const hashedNonce = await Crypto.digestStringAsync(
-        Crypto.CryptoDigestAlgorithm.SHA256,
-        nonce,
-      );
-
-      const appleCredential = await AppleAuthentication.signInAsync({
-        requestedScopes: [
-          AppleAuthentication.AppleAuthenticationScope.FULL_NAME,
-          AppleAuthentication.AppleAuthenticationScope.EMAIL,
-        ],
-        state: csrf,
-        nonce: hashedNonce,
-      });
-
-      const {identityToken} = appleCredential;
-
-      if (identityToken) {
-        const mutationConfig = {
-          variables: {
-            accessToken: identityToken,
-          },
-          onCompleted: (response: UserSignInAppleMutationResponse) => {
-            const {token, user} = response.signInWithApple as AuthPayload;
-
-            AsyncStorage.setItem('token', token);
-
-            createNotificationIfPushTokenExists();
-
-            setUser(user as User);
-          },
-          onError: (err: Error) => {
-            showAlertForError(err);
-          },
-        };
-
-        commitApple(mutationConfig);
-      }
-    } catch (e) {
-      if (e.code === 'ERR_CANCELED') {
-        // handle that the user canceled the sign-in flow
-      } else
-        Platform.select({
-          // @ts-ignore
-          // eslint-disable-next-line no-alert
-          web: alert(`Apple Login Error: ${e.code} - ${e.message}`),
-          default: Alert.alert(`Apple Login Error: ${e.code} - ${e.message}`),
-        });
-    } finally {
-      setSigningInApple(false);
-    }
-  };
-
-  useEffect(() => {
-    // console.log('appOwnership', Constants.appOwnership);
-  }, []);
+const SignIn: FC<Props> = ({
+  email,
+  errorEmail,
+  password,
+  errorPassword,
+  isSigningIn,
+  isSigningInWithApple,
+  onAppIconPressed,
+  onSignInPressed,
+  onSignUpPressed,
+  onFindPwPressed,
+  onSignInWithApplePressed,
+  onUserCreated,
+  onAgreementPressed,
+  onPrivacyPressed,
+  onChangeEmailText,
+  onChangePasswordText,
+}): ReactElement => {
+  const {theme, themeType} = useTheme();
 
   const logoSize = 80;
   const logoTransformAnimValue = useValue(0);
@@ -357,7 +208,7 @@ function SignIn(props: Props): ReactElement {
       <StyledScrollView>
         <AnimatedTouchableOpacity
           testID="theme-test"
-          onPress={(): void => changeThemeType()}
+          onPress={onAppIconPressed}
           style={{
             zIndex: 15,
             position: 'absolute',
@@ -393,12 +244,9 @@ function SignIn(props: Props): ReactElement {
             placeholderTextColor={theme.placeholder}
             placeholder="hello@example.com"
             value={email}
-            onChangeText={(text: string): void => {
-              setEmail(text);
-              setErrorEmail('');
-            }}
+            onChangeText={onChangeEmailText}
             errorText={errorEmail}
-            onSubmitEditing={signIn}
+            onSubmitEditing={onSignInPressed}
           />
           <EditText
             testID="input-password"
@@ -418,18 +266,15 @@ function SignIn(props: Props): ReactElement {
             placeholder="******"
             placeholderTextColor={theme.placeholder}
             value={password}
-            onChangeText={(text: string): void => {
-              setPassword(text);
-              setErrorPassword('');
-            }}
+            onChangeText={onChangePasswordText}
             errorText={errorPassword}
-            onSubmitEditing={signIn}
+            onSubmitEditing={onSignInPressed}
             secureTextEntry={true}
           />
           <ButtonWrapper>
             <Button
               testID="btn-sign-up"
-              onPress={goToSignUp}
+              onPress={onSignUpPressed}
               style={{flex: 1}}
               styles={{
                 container: {
@@ -455,8 +300,8 @@ function SignIn(props: Props): ReactElement {
             <View style={{width: 12}} />
             <Button
               testID="btn-sign-in"
-              loading={isInFlight}
-              onPress={signIn}
+              loading={isSigningIn}
+              onPress={onSignInPressed}
               style={{
                 flex: 1,
               }}
@@ -483,7 +328,7 @@ function SignIn(props: Props): ReactElement {
               text={getString('LOGIN')}
             />
           </ButtonWrapper>
-          <FindPwTouchOpacity testID="btn-find-pw" onPress={goToFindPw}>
+          <FindPwTouchOpacity testID="btn-find-pw" onPress={onFindPwPressed}>
             <FindPwText>{getString('FORGOT_PW')}</FindPwText>
           </FindPwTouchOpacity>
           <SocialButtonWrapper>
@@ -511,9 +356,9 @@ function SignIn(props: Props): ReactElement {
                       <SvgApple width={18} height={18} fill={theme.appleIcon} />
                     </View>
                   }
-                  loading={signingInApple || isAppleInFlight}
+                  loading={isSigningInWithApple}
                   indicatorColor={theme.primary}
-                  onPress={appleLogin}
+                  onPress={onSignInWithApplePressed}
                   text={getString('SIGN_IN_WITH_APPLE')}
                 />
               ),
@@ -522,20 +367,14 @@ function SignIn(props: Props): ReactElement {
               svgIcon={
                 <SvgFacebook width={18} height={18} fill={theme.facebookIcon} />
               }
-              onUserCreated={(user?: User): void => {
-                createNotificationIfPushTokenExists();
-                setUser?.(user);
-              }}
+              onUserCreated={onUserCreated}
               socialProvider={'facebook'}
             />
             <SocialSignInButton
               svgIcon={
                 <SvgGoogle width={20} height={20} fill={theme.googleIcon} />
               }
-              onUserCreated={(user?: User): void => {
-                createNotificationIfPushTokenExists();
-                setUser?.(user);
-              }}
+              onUserCreated={onUserCreated}
               socialProvider="google"
             />
           </SocialButtonWrapper>
@@ -543,17 +382,13 @@ function SignIn(props: Props): ReactElement {
             <StyledAgreementText>{getString('AGREEMENT1')}</StyledAgreementText>
             <StyledAgreementLinedText
               testID="btn-terms"
-              onPress={(): void =>
-                goToWebView('https://legacy.dooboolab.com/termsofservice')
-              }>
+              onPress={onAgreementPressed}>
               {getString('AGREEMENT2')}
             </StyledAgreementLinedText>
             <StyledAgreementText>{getString('AGREEMENT3')}</StyledAgreementText>
             <StyledAgreementLinedText
               testID="btn-privacy"
-              onPress={(): void =>
-                goToWebView('https://legacy.dooboolab.com/privacyandpolicy')
-              }>
+              onPress={onPrivacyPressed}>
               {getString('AGREEMENT4')}
             </StyledAgreementLinedText>
             <StyledAgreementText>{getString('AGREEMENT5')}</StyledAgreementText>
@@ -562,6 +397,6 @@ function SignIn(props: Props): ReactElement {
       </StyledScrollView>
     </Container>
   );
-}
+};
 
 export default SignIn;
